Change condition expression to disable services
The condition that checks the properties to enable/disable services
(`logsearch.logs.enabled` and `logsearch.metrics.enabled`) do not
work as expected.

It only checks if the property is defined with the `if_p`, and that always
evaluate to true regardless the value of the property, even when
the value is `false:FalseClass`. Because that, there is no way to
actually disable the services.

Instead of checking if the property is defined, we must check if the
value of the property evals to true.